refactor(b-ticket): move independent ticket files to ticket directory
