+++ backend/src/main/java/com/kett/TicketSystem/ticket/domain/exceptions/NoTicketFoundException.java
@@ -1,4 +1,4 @@
-package com.kett.TicketSystem.project.domain.exceptions;
+package com.kett.TicketSystem.ticket.domain.exceptions;
 
 public class NoTicketFoundException extends RuntimeException {
     public NoTicketFoundException(String message) {
