+++ backend/src/main/java/com/kett/TicketSystem/ticket/domain/exceptions/TicketException.java
@@ -1,4 +1,4 @@
-package com.kett.TicketSystem.project.domain.exceptions;
+package com.kett.TicketSystem.ticket.domain.exceptions;
 
 public class TicketException extends RuntimeException {
     public TicketException(String message) {
